feat: add SenPlayer support and localization updates
- add SenPlayer as a new external player option on MacOS
- add localization strings for SenPlayer launch settings in multiple languages.
- update player interaction logic to accommodate SenPlayer alongside Infuse.
