+++ public/lang/be.js
@@ -157,6 +157,11 @@ export default {
     settings_infuse_launch_ask: 'Пытаць пры запуску',
     settings_infuse_launch_play: 'Гуляць',
     settings_infuse_launch_save_and_play: 'Захаваць і гуляць',
+    settings_senplayer_launch_mode: 'Запуск у SenPlayer',
+    settings_senplayer_launch_mode_descr: 'Што рабіць пры адкрыцці відэа ў SenPlayer. Для торэнтаў заўжды адразу прайграванне',
+    settings_senplayer_launch_ask: 'Пытаць пры запуску',
+    settings_senplayer_launch_play: 'Гуляць',
+    settings_senplayer_launch_save_and_play: 'Захаваць і гуляць',
     settings_player_iptv_type: 'Тып плэера для IPTV',
     settings_player_iptv_type_descr: 'Якім плэерам прайграваць IPTV каналы',
     settings_player_reset: 'Скінуць плэер па змаўчанні',
@@ -333,6 +338,7 @@ export default {
     title_last: 'Апошняя',
     title_action: 'Дзеянне',
     title_action_infuse: 'Дзеянне Infuse',
+    title_action_senplayer: 'Дзеянне SenPlayer',
     title_producer: 'Рэжысёр',
     title_collection: 'Калекцыя',
     title_recomendations: 'Рэкамендацыі',
