+++ public/lang/en.js
@@ -158,6 +158,11 @@ export default {
     settings_infuse_launch_ask: 'Ask on launch',
     settings_infuse_launch_play: 'Play',
     settings_infuse_launch_save_and_play: 'Save and play',
+    settings_senplayer_launch_mode: 'Launch in SenPlayer',
+    settings_senplayer_launch_mode_descr: 'What to do when opening video in SenPlayer. Torrents always play immediately',
+    settings_senplayer_launch_ask: 'Ask on launch',
+    settings_senplayer_launch_play: 'Play',
+    settings_senplayer_launch_save_and_play: 'Save and play',
     settings_player_iptv_type: 'Player type for IPTV',
     settings_player_iptv_type_descr: 'Which player to play IPTV channels',
     settings_player_reset: 'Reset default player',
@@ -333,6 +338,7 @@ export default {
     title_last: 'Last',
     title_action: 'Action',
     title_action_infuse: 'Infuse action',
+    title_action_senplayer: 'SenPlayer action',
     title_producer: 'Producer',
     title_collection: 'Collection',
     title_recomendations: 'Recommendations',
