+++ public/lang/pl.js
@@ -193,6 +193,11 @@ export default {
     settings_infuse_launch_ask:  'Pytaj przy uruchomieniu',
     settings_infuse_launch_play:  'Odtwórz',
     settings_infuse_launch_save_and_play:  'Zapisz i odtwórz',
+    settings_senplayer_launch_mode: 'Uruchamianie w SenPlayer',
+    settings_senplayer_launch_mode_descr: 'Co robić przy otwieraniu wideo w SenPlayer. Torrenty zawsze odtwarzane od razu',
+    settings_senplayer_launch_ask: '#{settings_infuse_launch_ask}',
+    settings_senplayer_launch_play: '#{settings_infuse_launch_play}',
+    settings_senplayer_launch_save_and_play: '#{settings_infuse_launch_save_and_play}',
     settings_player_iptv_type:  'Typ odtwarzacza IPTV',
     settings_player_iptv_type_descr:  'Którego odtwarzacza używać do kanałów IPTV',
     settings_player_torrent_type:  'Typ odtwarzacza torrentów',
@@ -407,6 +412,7 @@ export default {
     title_last:  'Ostatnie',
     title_action:  'Akcja',
     title_action_infuse:  'Akcja Infuse',
+    title_action_senplayer: 'Akcja SenPlayer',
     title_producer:  'Reżyser',
     title_collection:  'Kolekcja',
     title_recomendations:  'Rekomendacje',
