+++ public/lang/ro.js
@@ -158,6 +158,11 @@ export default {
     settings_infuse_launch_ask: 'Întreabă la lansare',
     settings_infuse_launch_play: 'Redă',
     settings_infuse_launch_save_and_play: 'Salvează și redă',
+    settings_senplayer_launch_mode: 'Lansare în SenPlayer',
+    settings_senplayer_launch_mode_descr: 'Ce să facă la deschiderea videoclipului în SenPlayer. Torrentele redau imediat',
+    settings_senplayer_launch_ask: 'Întreabă la lansare',
+    settings_senplayer_launch_play: 'Redă',
+    settings_senplayer_launch_save_and_play: 'Salvează și redă',
     settings_player_iptv_type: 'Tip player pentru IPTV',
     settings_player_iptv_type_descr: 'Ce player să fie folosit pentru canale IPTV',
     settings_player_reset: 'Resetează player implicit',
@@ -333,6 +338,7 @@ export default {
     title_last: 'Ultimele',
     title_action: 'Acțiune',
     title_action_infuse: 'Acțiune Infuse',
+    title_action_senplayer: 'Acțiune SenPlayer',
     title_producer: 'Producător',
     title_collection: 'Colecție',
     title_recomendations: 'Recomandări',
