+++ public/lang/uk.js
@@ -148,6 +148,11 @@ export default {
     settings_infuse_launch_ask: 'Питати при запуску',
     settings_infuse_launch_play: 'Грати',
     settings_infuse_launch_save_and_play: 'Зберегти і грати',
+    settings_senplayer_launch_mode: 'Запуск в SenPlayer',
+    settings_senplayer_launch_mode_descr: 'Що робити при відкритті відео в SenPlayer. Для торрентів завжди одразу відтворення',
+    settings_senplayer_launch_ask: 'Питати при запуску',
+    settings_senplayer_launch_play: 'Грати',
+    settings_senplayer_launch_save_and_play: 'Зберегти і грати',
     settings_player_iptv_type: 'Тип плеєра для IPTV',
     settings_player_iptv_type_descr: 'Яким плеєром відтворювати IPTV канали',
     settings_player_reset: 'Скинути програвач за замовчуванням',
@@ -311,6 +316,7 @@ export default {
     title_last: 'Остання',
     title_action: 'Дія',
     title_action_infuse: 'Дія Infuse',
+    title_action_senplayer: 'Дія SenPlayer',
     title_producer: 'Режисер',
     title_collection: 'Колекція',
     title_recomendations: 'Рекомендації',
