+++ src/interaction/settings/params.js
@@ -102,6 +102,7 @@ function init(){
             'inner': '#{settings_param_player_inner}',
             'iina': 'IINA',
             'infuse': 'Infuse',
+            'senplayer': 'SenPlayer',
             'mpv': 'MPV',
             'nplayer': 'nPlayer',
             'tracyplayer': 'TracyPlayer',		
@@ -111,6 +112,7 @@ function init(){
             'inner': '#{settings_param_player_inner}',
             'iina': 'IINA',
             'infuse': 'Infuse',
+            'senplayer': 'SenPlayer',
             'mpv': 'MPV',
             'nplayer': 'nPlayer',
             'tracyplayer': 'TracyPlayer',		
@@ -120,6 +122,7 @@ function init(){
             'inner': '#{settings_param_player_inner}',
             'iina': 'IINA',
             'infuse': 'Infuse',
+            'senplayer': 'SenPlayer',
             'mpv': 'MPV',
             'nplayer': 'nPlayer',
             'tracyplayer': 'TracyPlayer',		
@@ -483,19 +486,20 @@ function update(elem,elems,elems_html){
         listener.send('update_scroll_position')
     }
 
-    updateInfuseLaunchVisibility(elems_html)
+    updateXCallbackLaunchVisibility(elems_html)
 }
 
-function isInfusePlayerSelected(){
+function isXCallbackPlayerSelected(id){
     return ['player', 'player_iptv', 'player_torrent'].some((name)=>{
-        return Storage.field(name) === 'infuse'
+        return Storage.field(name) === id
     })
 }
 
-function updateInfuseLaunchVisibility(elems_html){
+function updateXCallbackLaunchVisibility(elems_html){
     if(!elems_html || !elems_html.length) return
 
-    elems_html.find('[data-infuse-launch]').toggleClass('hide', !isInfusePlayerSelected())
+    elems_html.find('[data-infuse-launch]').toggleClass('hide', !isXCallbackPlayerSelected('infuse'))
+    elems_html.find('[data-senplayer-launch]').toggleClass('hide', !isXCallbackPlayerSelected('senplayer'))
 }
 
 /**
@@ -625,6 +629,12 @@ select('infuse_launch_mode', {
     'ask': '#{settings_infuse_launch_ask}',
     'play': '#{settings_infuse_launch_play}',
     'save_and_play': '#{settings_infuse_launch_save_and_play}',
+}, 'play')
+
+select('senplayer_launch_mode', {
+    'ask': '#{settings_senplayer_launch_ask}',
+    'play': '#{settings_senplayer_launch_play}',
+    'save_and_play': '#{settings_senplayer_launch_save_and_play}',
 }, 'play')
 
 select('player_scale_method',{
