+++ src/lang/bg.js
@@ -157,6 +157,11 @@ export default {
     settings_infuse_launch_ask: 'Питай при стартиране',
     settings_infuse_launch_play: 'Пусни',
     settings_infuse_launch_save_and_play: 'Запази и пусни',
+    settings_senplayer_launch_mode: 'Стартиране в SenPlayer',
+    settings_senplayer_launch_mode_descr: 'Какво да прави при отваряне на видео в SenPlayer. За торенти винаги веднага възпроизвеждане',
+    settings_senplayer_launch_ask: 'Питай при стартиране',
+    settings_senplayer_launch_play: 'Пусни',
+    settings_senplayer_launch_save_and_play: 'Запази и пусни',
     settings_player_iptv_type: 'Тип плеър за IPTV',
     settings_player_iptv_type_descr: 'Кой плейър да се използва за IPTV канали',
     settings_player_reset: 'Нулирай плейъра по подразбиране',
@@ -332,6 +337,7 @@ export default {
     title_last: 'Последен',
     title_action: 'Действие',
     title_action_infuse: 'Действие Infuse',
+    title_action_senplayer: 'Действие SenPlayer',
     title_producer: 'Продуцент',
     title_collection: 'Колекция',
     title_recomendations: 'Препоръки',
