+++ src/lang/cs.js
@@ -185,11 +185,16 @@ export default {
 
     settings_player_type: "Typ přehrávače",
     settings_player_type_descr: "Jakým přehrávačem přehrávat video",
-    settings_infuse_launch_mode: "Spuštění v Infuse",
-    settings_infuse_launch_mode_descr: "Co dělat při otevření videa v Infuse. U torrentů vždy okamžité přehrání",
+    settings_infuse_launch_mode: 'Spuštění v Infuse',
+    settings_infuse_launch_mode_descr: 'Co dělat při otevření videa v Infuse. U torrentů vždy okamžité přehrání',
     settings_infuse_launch_ask: "Zeptat se při spuštění",
     settings_infuse_launch_play: "Přehrát",
     settings_infuse_launch_save_and_play: "Uložit a přehrát",
+    settings_senplayer_launch_mode: 'Spuštění v SenPlayer',
+    settings_senplayer_launch_mode_descr: 'Co dělat při otevření videa v SenPlayer. U torrentů vždy okamžité přehrání',
+    settings_senplayer_launch_ask: "Zeptat se při spuštění",
+    settings_senplayer_launch_play: "Přehrát",
+    settings_senplayer_launch_save_and_play: "Uložit a přehrát",
     settings_player_iptv_type: "Typ přehrávače pro IPTV",
     settings_player_iptv_type_descr: "Jakým přehrávačem přehrávat IPTV kanály",
     settings_player_torrent_type: "Typ přehrávače pro torrenty",
@@ -400,6 +405,7 @@ export default {
     title_last: "Poslední",
     title_action: "Akce",
     title_action_infuse: "Akce Infuse",
+    title_action_senplayer: 'Akce SenPlayer',
     title_producer: "Režisér",
     title_collection: "Kolekce",
     title_recomendations: "Doporučení",
