+++ src/lang/fr.js
@@ -158,6 +158,11 @@ export default {
     settings_infuse_launch_ask: 'Demander au lancement',
     settings_infuse_launch_play: 'Lire',
     settings_infuse_launch_save_and_play: 'Enregistrer et lire',
+    settings_senplayer_launch_mode: 'Lancement dans SenPlayer',
+    settings_senplayer_launch_mode_descr: 'Action à l\'ouverture d\'une vidéo dans SenPlayer. Les torrents sont toujours lus immédiatement',
+    settings_senplayer_launch_ask: 'Demander au lancement',
+    settings_senplayer_launch_play: 'Lire',
+    settings_senplayer_launch_save_and_play: 'Enregistrer et lire',
     settings_player_iptv_type: 'Lecteur pour l\'IPTV',
     settings_player_iptv_type_descr: 'Lecteur utilisé pour les chaînes IPTV',
     settings_player_reset: 'Réinitialiser le lecteur',
@@ -333,6 +338,7 @@ export default {
     title_last: 'Dernier',
     title_action: 'Action',
     title_action_infuse: 'Action Infuse',
+    title_action_senplayer: 'Action SenPlayer',
     title_producer: 'Producteur',
     title_collection: 'Collection',
     title_recomendations: 'Recommandations',
