+++ src/lang/he.js
@@ -144,6 +144,11 @@ export default {
     settings_infuse_launch_ask: 'שאל בהפעלה',
     settings_infuse_launch_play: 'השמע',
     settings_infuse_launch_save_and_play: 'שמור והשמע',
+    settings_senplayer_launch_mode: 'הפעלה ב-SenPlayer',
+    settings_senplayer_launch_mode_descr: 'מה לעשות בפתיחת וידאו ב-SenPlayer. טורנטים תמיד מתחילים מיד',
+    settings_senplayer_launch_ask: 'שאל בהפעלה',
+    settings_senplayer_launch_play: 'השמע',
+    settings_senplayer_launch_save_and_play: 'שמור והשמע',
     settings_player_iptv_type: 'IPTV סוג נגן עבור',
     settings_player_iptv_type_descr: 'IPTV איזה נגן להפעיל בערוצי',
     settings_player_reset: 'אפס נגן ברירת מחדל',
@@ -307,6 +312,7 @@ export default {
     title_last: 'אחרון',
     title_action: 'פעולה',
     title_action_infuse: 'פעולת Infuse',
+    title_action_senplayer: 'פעולת SenPlayer',
     title_producer: 'מפיק',
     title_collection: 'אוסף',
     title_recomendations: 'המלצות',
