+++ src/lang/pt.js
@@ -156,6 +156,11 @@ export default {
     settings_infuse_launch_ask: 'Perguntar ao iniciar',
     settings_infuse_launch_play: 'Reproduzir',
     settings_infuse_launch_save_and_play: 'Guardar e reproduzir',
+    settings_senplayer_launch_mode: 'Iniciar no SenPlayer',
+    settings_senplayer_launch_mode_descr: 'O que fazer ao abrir vídeo no SenPlayer. Torrents sempre reproduzem imediatamente',
+    settings_senplayer_launch_ask: 'Perguntar ao iniciar',
+    settings_senplayer_launch_play: 'Reproduzir',
+    settings_senplayer_launch_save_and_play: 'Guardar e reproduzir',
     settings_player_iptv_type: 'Tipo de player para IPTV',
     settings_player_iptv_type_descr: 'Qual player reproduzir canais de IPTV',
     settings_player_reset: 'Redefinir player padrão',
@@ -331,6 +336,7 @@ export default {
     title_last: 'Último',
     title_action: 'Ação',
     title_action_infuse: 'Ação Infuse',
+    title_action_senplayer: 'Ação SenPlayer',
     title_producer: 'Produtor',
     title_collection: 'Coleção',
     title_recomendations: 'Recomendações',
