+++ src/lang/ru.js
@@ -193,6 +193,11 @@ export default {
     settings_infuse_launch_ask: 'Спрашивать при запуске',
     settings_infuse_launch_play: 'Играть',
     settings_infuse_launch_save_and_play: 'Сохранить и играть',
+    settings_senplayer_launch_mode: 'Запуск в SenPlayer',
+    settings_senplayer_launch_mode_descr: 'Что делать при открытии видео в SenPlayer. Для торрентов всегда сразу воспроизведение',
+    settings_senplayer_launch_ask: 'Спрашивать при запуске',
+    settings_senplayer_launch_play: 'Играть',
+    settings_senplayer_launch_save_and_play: 'Сохранить и играть',
     settings_player_iptv_type: 'Тип плеера для IPTV',
     settings_player_iptv_type_descr: 'Каким плеером воспроизводить IPTV-каналы',
     settings_player_torrent_type: 'Тип плеера для торрентов',
@@ -410,6 +415,7 @@ export default {
     title_last: 'Последняя',
     title_action: 'Действие',
     title_action_infuse: 'Действие Infuse',
+    title_action_senplayer: 'Действие SenPlayer',
     title_producer: 'Режиссер',
     title_collection: 'Коллекция',
     title_recomendations: 'Рекомендации',
