+++ src/lang/zh.js
@@ -146,6 +146,11 @@ export default {
     settings_infuse_launch_ask: '启动时询问',
     settings_infuse_launch_play: '播放',
     settings_infuse_launch_save_and_play: '保存并播放',
+    settings_senplayer_launch_mode: 'SenPlayer 启动方式',
+    settings_senplayer_launch_mode_descr: '在 SenPlayer 中打开视频时的操作。种子始终立即播放',
+    settings_senplayer_launch_ask: '启动时询问',
+    settings_senplayer_launch_play: '播放',
+    settings_senplayer_launch_save_and_play: '保存并播放',
     settings_player_iptv_type: 'IPTV播放器类型',
     settings_player_iptv_type_descr: '哪个播放器播放IPTV频道',
     settings_player_reset: '重置默认播放器',
@@ -311,6 +316,7 @@ export default {
     title_last: '最后',
     title_action: '动作',
     title_action_infuse: 'Infuse 操作',
+    title_action_senplayer: 'SenPlayer 操作',
     title_producer: '制片人',
     title_collection: '合集',
     title_recomendations: '推荐',
